Experience - Fix fresnel shader without point light

+++ src/components/WebGLExperience/meshes/FishGroup/index.js
@@ -73,8 +73,11 @@ class FishGroup extends THREE.Group {
 
       if( index > -1 ) {
         this.fishes.splice( index, 1 );
-        fish.pointLightTl.kill();
-        fish.pointLight.intensity = 0;
+
+        if( fish.species === 'lanternFish' ) {
+          fish.pointLightTl.kill();
+          fish.pointLight.intensity = 0;
+        }
 
         this.remove( fish.modelObject );
       }
